refactor(header-chain): split transition planner into phases
Make phase ordering and projected-state coupling explicit so event domains and invariant boundaries can evolve independently.
