+++ crates/zakura-header-chain/src/transition/mod.rs
@@ -13,12 +13,12 @@ pub use engine::{
     CommittedTransitionError, DurableTransitionFacts, EngineHydrationError, EngineTransition,
     HeaderChainEngine,
 };
-pub(crate) use invariants::verify_plan;
-#[cfg(any(test, feature = "fuzz-impl"))]
-pub(crate) use invariants::verify_plan_production;
+pub(crate) use invariants::verify_candidate;
 pub use invariants::InvariantViolation;
+#[cfg(any(test, feature = "fuzz-impl"))]
+pub(crate) use invariants::{verify_plan, verify_plan_production};
 pub use planner::TransitionFailure;
-pub(crate) use planner::TransitionPlan;
+pub(crate) use planner::{PlanCandidate, TransitionPlan};
 pub use recovery::{
     audit_store, audit_store_at, audit_store_for_trust_anchor_update, AuditViolation,
     RecoveryFailure, RecoveryPlan, RecoveryRepair, StoreAuditRead, ValidationContextRecord,
