+++ crates/zakura-header-chain/src/transition/planner/event_effects/deferred_time.rs
@@ -1,0 +1,31 @@
+//! Deferred local-time header reevaluation.
+
+use crate::graph::HeaderGraphView;
+use crate::{HeaderValidationState, TransitionFailure};
+
+use super::super::projected_state::ProjectedTransitionState;
+use super::ApplyEventContext;
+
+/// Promote every due deferred header to valid using the authoritative clock.
+pub(super) fn apply(
+    projected: &mut ProjectedTransitionState<'_>,
+    event_context: &ApplyEventContext<'_>,
+) -> Result<(), TransitionFailure> {
+    let due: Vec<_> = projected
+        .graph()
+        .view_header_nodes()
+        .into_iter()
+        .filter_map(|node| match node.validation {
+            HeaderValidationState::DeferredUntil(until)
+                if until <= event_context.transition.clock.now() =>
+            {
+                Some(node.hash)
+            }
+            _ => None,
+        })
+        .collect();
+    for hash in due {
+        projected.set_header_validation_state(hash, HeaderValidationState::Valid)?;
+    }
+    Ok(())
+}
