+++ crates/zakura-header-chain/src/transition/planner/settlement.rs
@@ -1,0 +1,214 @@
+//! Finality, retention, and projection settlement after event evidence.
+
+use std::borrow::Cow;
+
+use crate::graph::HeaderGraphView;
+use crate::retention::RetentionPlan;
+use crate::{
+    EngineMetadata, EngineMode, EngineSnapshot, FinalityRecord, FinalitySource, Frontier,
+    HeaderChainEngine, TransitionCause, TransitionContext, TransitionEvent, TransitionFailure,
+};
+
+use super::admission::HeaderInsertionRebase;
+use super::projected_state::{path, ProjectedTransitionState, SettledProjectedState};
+
+/// Inputs required to settle finality and retention after event evidence.
+pub(super) struct SettlementInputs<'engine, 'ctx> {
+    pub(super) engine: &'engine HeaderChainEngine,
+    pub(super) projected: ProjectedTransitionState<'engine>,
+    pub(super) metadata: EngineMetadata,
+    pub(super) before: &'ctx EngineSnapshot,
+    pub(super) event: &'ctx TransitionEvent,
+    pub(super) header_rebase: HeaderInsertionRebase,
+    pub(super) context: &'ctx TransitionContext<'ctx>,
+    pub(super) old_selected: &'engine [Frontier],
+}
+
+/// Settled projections ready for write-set assembly.
+pub(super) struct SettledTransition<'a> {
+    /// Projected graph after finality and retention.
+    pub(super) projected: SettledProjectedState<'a>,
+    /// Selected projection after settlement.
+    pub(super) selected: Cow<'a, [Frontier]>,
+    /// Optional finality record to append.
+    pub(super) finality_append: Option<FinalityRecord>,
+    /// Retention outcome.
+    pub(super) retention: RetentionPlan,
+    /// Classified transition cause.
+    pub(super) cause: TransitionCause,
+    /// Updated metadata carrying work-origin and alarm side effects.
+    pub(super) metadata: EngineMetadata,
+}
+
+/// Outcome of finality and retention settlement.
+pub(super) enum FinalityRetentionOutcome<'a> {
+    /// Settlement completed and write assembly may proceed.
+    Settled(Box<SettledTransition<'a>>),
+    /// Protected paths made limits unenforceable; discard event effects.
+    ResourceStalled,
+}
+
+/// Decide and apply finality, enforce retention, and settle projections.
+pub(super) fn derive_finality_and_retention<'engine, 'ctx>(
+    inputs: SettlementInputs<'engine, 'ctx>,
+) -> Result<FinalityRetentionOutcome<'engine>, TransitionFailure> {
+    let SettlementInputs {
+        engine,
+        mut projected,
+        mut metadata,
+        before,
+        event,
+        header_rebase,
+        context,
+        old_selected,
+    } = inputs;
+    let work_rebased = projected.work_coordinates_rebased();
+    if work_rebased {
+        metadata.work_origin = projected.graph().view_finalized_frontier();
+    }
+    projected.refresh_verified_after_operator_change()?;
+
+    let (mut header_best, _) = projected.graph().view_select_best_header_chain()?;
+    let full_state_finalized = match event {
+        TransitionEvent::FullStateFinalized(event) => {
+            Some((event.new_finalized, event.full_state_transition_id))
+        }
+        TransitionEvent::VerifiedChainChanged(event)
+            if event.cause == crate::VerifiedChangeCause::CheckpointFinalizedGrow =>
+        {
+            event.new_path.last().map(|header| {
+                (
+                    Frontier::new(header.height, header.hash),
+                    event.full_state_transition_id,
+                )
+            })
+        }
+        _ => None,
+    };
+
+    let mut finality = None;
+    if let Some((new_finalized, evidence)) = full_state_finalized {
+        if new_finalized.height < before.frontiers.finalized.height {
+            return Err(TransitionFailure::InvalidEvidence("finality retreated"));
+        }
+        if !projected.verified().contains(&new_finalized) {
+            return Err(TransitionFailure::InvalidEvidence(
+                "integrated finality is not on the verified projection",
+            ));
+        }
+        finality = Some((new_finalized, FinalitySource::FullState { evidence }));
+    } else if context.config.mode == EngineMode::HeadersOnly {
+        let depth = context.config.limits.local_finality_depth.get();
+        if header_best
+            .height
+            .0
+            .saturating_sub(projected.graph().view_finalized_frontier().height.0)
+            > depth
+        {
+            let height = zakura_chain::block::Height(header_best.height.0 - depth);
+            let new_finalized = projected
+                .graph()
+                .view_header_ancestor(header_best.hash, height)?
+                .ok_or(TransitionFailure::InvalidEvidence(
+                    "selected ancestry is incomplete",
+                ))?;
+            finality = Some((
+                new_finalized,
+                FinalitySource::HeadersOnlyDepth {
+                    selected_tip: header_best,
+                },
+            ));
+        }
+    }
+
+    let mut cause = if work_rebased || header_rebase == HeaderInsertionRebase::Rebased {
+        TransitionCause::HeaderWorkRebased
+    } else if matches!(
+        event,
+        TransitionEvent::VerifiedChainChanged(ref event)
+            if event.cause == crate::VerifiedChangeCause::CheckpointFinalizedGrow
+    ) {
+        TransitionCause::CheckpointFinality
+    } else if matches!(event, TransitionEvent::AuxEvidence(_)) {
+        TransitionCause::AuxAuthentication
+    } else {
+        TransitionCause::Event
+    };
+
+    let mut finality_append = None;
+    if let Some((new_finalized, source)) = finality {
+        if new_finalized != projected.graph().view_finalized_frontier() {
+            let previous = projected.graph().view_finalized_frontier();
+            let epoch = metadata.finality_epoch.checked_next()?;
+            projected.advance_finality(new_finalized)?;
+            finality_append = Some(FinalityRecord {
+                previous,
+                current: new_finalized,
+                source,
+                epoch,
+            });
+            header_best = projected.graph().view_select_best_header_chain()?.0;
+            if matches!(source, FinalitySource::HeadersOnlyDepth { .. }) {
+                cause = TransitionCause::HeadersOnlyFinality;
+            }
+        }
+    }
+
+    if context.config.mode == EngineMode::HeadersOnly {
+        projected.force_headers_only_verified();
+    }
+
+    let retention = projected.enforce_retention(
+        header_best,
+        context.retention_references.iter().copied(),
+        context.config.limits,
+    )?;
+    if retention.admission_refused {
+        return Ok(FinalityRetentionOutcome::ResourceStalled);
+    }
+
+    header_best = projected.graph().view_select_best_header_chain()?.0;
+    let selected = if cause == TransitionCause::CheckpointFinality
+        && header_best == before.frontiers.header_best
+        && finality_append.is_some_and(|record| {
+            old_selected
+                .binary_search_by_key(&record.current.height, |frontier| frontier.height)
+                .ok()
+                .is_some_and(|index| old_selected[index] == record.current)
+        }) {
+        let Some(record) = finality_append else {
+            return Err(TransitionFailure::InvalidEvidence(
+                "checkpoint finality has no finality record",
+            ));
+        };
+        let index = old_selected
+            .binary_search_by_key(&record.current.height, |frontier| frontier.height)
+            .map_err(|_| {
+                TransitionFailure::InvalidEvidence(
+                    "checkpoint finality left the selected projection",
+                )
+            })?;
+        Cow::Borrowed(&old_selected[index..])
+    } else {
+        Cow::Owned(path(projected.graph(), header_best)?)
+    };
+
+    let projected = projected.finish_after_retention(engine)?;
+    Ok(FinalityRetentionOutcome::Settled(Box::new(
+        SettledTransition {
+            projected,
+            selected,
+            finality_append,
+            retention,
+            cause,
+            metadata,
+        },
+    )))
+}
+
+/// Apply a migrated-pin refutation alarm before settlement when durable facts authenticate it.
+pub(super) fn apply_migrated_pin_alarm(metadata: &mut EngineMetadata, pin: Option<Frontier>) {
+    if let Some(pin) = pin {
+        metadata.alarms.migrated_pin_refuted = Some(pin);
+    }
+}
