+++ crates/zakura-header-chain/src/transition/planner/tests/auxiliary.rs
@@ -1,3 +1,4 @@
+use super::super::projected_state::path;
 use super::*;
 use crate::{AuxDelta, InvariantViolation};
