+++ crates/zakura-header-chain/src/transition/planner/tests/operator.rs
@@ -1,3 +1,4 @@
+use super::super::projected_state::path;
 use super::*;
 
 #[test]
