+++ crates/zakura-header-chain/src/transition/planner/tests/selection.rs
@@ -1,3 +1,4 @@
+use super::super::projected_state::trim_projection;
 use super::*;
 
 #[test]
